refactor: move-valores
